Fix 1.8.7.
I'm not sure why, but aliasing `__send__` to `send`
on 1.8.7 doesn't work correctly. Defining the alias
manually does, for some reason.

While we're at it, it makes more sense to define
`__send__` as primary and `send` as secondary,
as it's more likely users will stub `send`
(e.g. for an `email` object).