Stop frozen object warnings from being issued multiple times per method.
This was being caused by the method double object not recording the fact
it had issued a warning. Given that we can't remove the method, and thus
don't tear down the method double, its state wasn't being modified when
the object was being reset.

This patch causes the method double object to record the fact that it
has issued the warning once, and should not continue to issue the
warning.

This also prevents the warning from being issued by the teardown of our
specs.