Remove need for `rescue NameError` on ruby 2.0.0.
This was needed because of a bug in our prepended
module detection. The logic was originally written
against ruby 2.1, where `object.singleton_class.ancestors`
always includes the singleton class. On ruby 2.0,
it only includes the singleton class if there is a
prepended module. If there is not a prepended module,
it's not in the ancestors array and our `take_while`
logic considered EVERY ancestor to be a prepended
module even though there were none.

Instead, we can find the prepended module by taking
all ancestors until the first class.