Rewrite cucumber features.
The old cukes were a mess:

- Lots of duplication and inconsistencies.
- Functioned poorly as documentation.
- Didn't highlight the new syntax well.
- Didn't really cover all of rspec-mocks' features.

The new ones were structurd specifically with documentation in mind.
The specs are meant more for regression coverage. We've limited the
cukes to just things people may actually want to read as part of
docs.

I've also reworked the test unit cuke as a minitest cuke.

For reference, here's the before/after from cucumber:

Before:

77 scenarios (77 passed)
276 steps (276 passed)
0m48.014s

After:

89 scenarios (89 passed)
328 steps (328 passed)
0m57.844s

Fixes #591.