Add more specific "does not implement" messages.
The unimplemented error message gives no indication of whether the
object is a class or instance double. This change passes that object to
the error generator so that it can create a more specific message
depending on whether it is an instance or class double. It defaults to
the original "<object> does not implement <method>" for anything else.

https://github.com/rspec/rspec-mocks/issues/838